fastapi-manager/pyproject.toml
[project]
name = "fastapi-manager"
version = "0.1.0"
description = "Media Manager API for downloads, sync, and organization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "yt-dlp>=2023.12.0",
    "guessit>=3.7.0",
    "tmdbsimple>=2.9.0",
    "mutagen>=1.47.0",
    "apscheduler>=3.10.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
